Support `db.getSync()` in get benchmark (#33)
This hypothetical method isn't supported anywhere in Level, except
in a local branch of `classic-level` that I'm putting aside for now.
This also isn't the right way to benchmark `getSync()`; it's only
mildly useful as a rough comparison to async `get()`. Which is
unsurprisingly a lot slower.

Long story short, I'm stashing some work.